+++ lib/features/player/logic/sessions_cleanup.g.dart
@@ -41,7 +41,7 @@ final class SessionsCleanupProvider
   }
 }
 
-String _$sessionsCleanupHash() => r'd3667b5842391e4350ef5a1b5c449bbfd6f6a7cc';
+String _$sessionsCleanupHash() => r'9c77f47038235c4fa038c475505527daf7ffbc7f';
 
 abstract class _$SessionsCleanup extends $Notifier<void> {
   void build();